+++ YACReaderLibrary/db/data_base_management.h
@@ -49,6 +49,7 @@ public:
     static QSqlDatabase loadDatabase(QString path);
     static QSqlDatabase loadDatabaseFromFile(QString path);
     static bool createTables(QSqlDatabase &database);
+    static bool createComicInfoTable(QSqlDatabase &database, QString tableName);
     static bool createV8Tables(QSqlDatabase &database);
 
     static void exportComicsInfo(QString source, QString dest);